+++ gwcelery/tests/test_tasks_p_astro_gstlal.py
@@ -1,12 +1,26 @@
 import json
+from unittest.mock import patch
+from urllib import request
 
 import numpy as np
 import pkg_resources
 import pytest
-from unittest.mock import patch
 
 from .. import app
 from ..tasks import p_astro_gstlal
+
+
+class MockResponse(object):
+    def read(self):
+        return json.dumps({
+            "counts_BNS": 2,
+            "counts_BBH": 10,
+            "counts_NSBH": 1,
+            "counts_Terrestrial": 4000,
+            "counts_MassGap": 1})
+
+    def close(self):
+        pass
 
 
 @pytest.fixture
@@ -61,7 +75,8 @@ def test_get_event_ln_likelihood_ratio_svd_endtime_mass(coinc_bytes_1):
     assert mass == pytest.approx(3.78, abs=0.1)
 
 
-def test_compute_p_astro_1(coinc_bytes_1, ranking_data_bytes):
+@patch.object(request, 'urlopen', return_value=MockResponse())
+def test_compute_p_astro_1(mock_url, coinc_bytes_1, ranking_data_bytes):
     """Test to call `compute_p_astro` on gracedb event G322589.
     m1 = 2.7, m2 = 1.0 solar mass for this event"""
     files = coinc_bytes_1, ranking_data_bytes
@@ -70,21 +85,34 @@ def test_compute_p_astro_1(coinc_bytes_1, ranking_data_bytes):
     assert p_astros['BNS'] == pytest.approx(1, abs=1e-2)
     assert p_astros['NSBH'] == pytest.approx(0, abs=1e-2)
     assert p_astros['BBH'] == pytest.approx(0, abs=1e-2)
-    assert p_astros['Terr'] == pytest.approx(0, abs=1e-2)
+    assert p_astros['MassGap'] == pytest.approx(0, abs=1e-2)
+    assert p_astros['Terrestrial'] == pytest.approx(0, abs=1e-2)
 
 
-def test_compute_p_astro_2(coinc_bytes_2, ranking_data_bytes):
+@patch.object(request, 'urlopen', return_value=MockResponse())
+def test_compute_p_astro_2(mock_url, coinc_bytes_2, ranking_data_bytes):
     """Test to call `compute_p_astro` on gracedb event G5351.
-    m1 = 4,2, m2 = 1.2 solar mass for this event. FAR = 1.9e-6, P_terr
+    m1 = 4.2, m2 = 1.2 solar mass for this event. FAR = 1.9e-6, P_terr
     has a high value."""
     files = coinc_bytes_2, ranking_data_bytes
     # FIXME graceid should be removed once fixed
     p_astros = json.loads(p_astro_gstlal.compute_p_astro(files))
-    assert p_astros['Terr'] > 0.85
+    assert p_astros['Terrestrial'] > 0.85
 
 
-def test_failing_compute_p_astro(coinc_bytes_3, ranking_data_bytes):
+@patch.object(request, 'urlopen', return_value=MockResponse())
+def test_failing_compute_p_astro(mock_url, coinc_bytes_3, ranking_data_bytes):
     """Test the case when p_astro computation fails"""
+    files = coinc_bytes_3, ranking_data_bytes
+    with patch('gwcelery.tasks.p_astro_other.compute_p_astro') as p:
+        p_astro_gstlal.compute_p_astro.s(files).delay()
+        p.assert_called_once()
+
+
+def test_handling_urlerror(coinc_bytes_3, ranking_data_bytes):
+    """Test case when the request to fetch mean values results in
+    a URLError if file is not found. p_astro_other should be called.
+    """
     files = coinc_bytes_3, ranking_data_bytes
     with patch('gwcelery.tasks.p_astro_other.compute_p_astro') as p:
         p_astro_gstlal.compute_p_astro.s(files).delay()